[loop-analysis] Organize mapping-to-isl code to a directory
