+++ include/loop-analysis/mapping-to-isl/fused-mapping-to-isl.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "loop-analysis/isl-ir.hpp"
+
+namespace analysis
+{
+
+std::map<LogicalBuffer, Occupancy>
+OccupanciesFromMapping(mapping::FusedMapping& mapping,
+                       const problem::FusedWorkload& workload);
+
+}; // namespace analysis
