+++ include/loop-analysis/temporal-analysis.hpp
@@ -12,12 +12,10 @@ struct BufTemporalReuseOpts
 
 struct TemporalReuseAnalysisInput
 {
-  const LogicalBuffer& logical_buffer;
   const Occupancy& occupancy;
   BufTemporalReuseOpts reuse_opts;
 
-  TemporalReuseAnalysisInput(const LogicalBuffer& logical_buffer,
-                             const Occupancy& occupancy,
+  TemporalReuseAnalysisInput(const Occupancy& occupancy,
                              BufTemporalReuseOpts reuse_opts);
 };
 
@@ -27,6 +25,15 @@ struct TemporalReuseAnalysisOutput
   Fill fill;
 };
 
+/**
+ * @brief Computes the required fill to satisfy the buffer occupancy.
+ * 
+ * If the buffer can `exploit_temporal_reuse`, then the fill will only consist
+ * of data not currently resident in buffer.
+ * 
+ * @see analysis::TemporalReuseAnalysisInput
+ * @see analysis::TemporalReuseAnalysisOutput
+ */
 TemporalReuseAnalysisOutput
 TemporalReuseAnalysis(TemporalReuseAnalysisInput input);
